feature: enchance endpoint security
`/roles` and `/division` can be access by admin only
`/employee` (PUT & DELETE) can only be access by same user or admin

+++ internal/app/employee/handler.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/Alterra-DataOn-Kelompok-5/employee-service/internal/dto"
 	"github.com/Alterra-DataOn-Kelompok-5/employee-service/internal/factory"
+	"github.com/Alterra-DataOn-Kelompok-5/employee-service/internal/pkg/enum"
 	"github.com/Alterra-DataOn-Kelompok-5/employee-service/internal/pkg/util"
 	pkgdto "github.com/Alterra-DataOn-Kelompok-5/employee-service/pkg/dto"
 	res "github.com/Alterra-DataOn-Kelompok-5/employee-service/pkg/util/response"
@@ -25,12 +26,10 @@ func NewHandler(f *factory.Factory) *handler {
 
 func (h *handler) Get(c echo.Context) error {
 	authHeader := c.Request().Header.Get("Authorization")
-	jwtClaims, err := util.ParseJWTToken(authHeader)
+	_, err := util.ParseJWTToken(authHeader)
 	if err != nil {
 		return res.ErrorBuilder(&res.ErrorConstant.Unauthorized, err).Send(c)
 	}
-
-	log.Println(jwtClaims)
 
 	payload := new(pkgdto.SearchGetRequest)
 	if err := c.Bind(payload); err != nil {
@@ -51,12 +50,10 @@ func (h *handler) Get(c echo.Context) error {
 
 func (h *handler) GetById(c echo.Context) error {
 	authHeader := c.Request().Header.Get("Authorization")
-	jwtClaims, err := util.ParseJWTToken(authHeader)
+	_, err := util.ParseJWTToken(authHeader)
 	if err != nil {
 		return res.ErrorBuilder(&res.ErrorConstant.Unauthorized, err).Send(c)
 	}
-
-	log.Println(jwtClaims)
 
 	payload := new(pkgdto.ByIDRequest)
 	if err := c.Bind(payload); err != nil {
@@ -75,20 +72,19 @@ func (h *handler) GetById(c echo.Context) error {
 }
 
 func (h *handler) UpdateById(c echo.Context) error {
-	authHeader := c.Request().Header.Get("Authorization")
-	jwtClaims, err := util.ParseJWTToken(authHeader)
-	if err != nil {
-		return res.ErrorBuilder(&res.ErrorConstant.Unauthorized, err).Send(c)
-	}
-
-	log.Println(jwtClaims)
-
 	payload := new(dto.UpdateEmployeeRequestBody)
 	if err := c.Bind(payload); err != nil {
 		return res.ErrorBuilder(&res.ErrorConstant.BadRequest, err).Send(c)
 	}
 	if err := c.Validate(payload); err != nil {
 		return res.ErrorBuilder(&res.ErrorConstant.Validation, err).Send(c)
+	}
+	authHeader := c.Request().Header.Get("Authorization")
+	jwtClaims, err := util.ParseJWTToken(authHeader)
+	isAdminOrSameUser := (jwtClaims.UserID == *payload.ID) || (jwtClaims.RoleID == uint(enum.Admin))
+	log.Println(isAdminOrSameUser)
+	if (err != nil) || !isAdminOrSameUser {
+		return res.ErrorBuilder(&res.ErrorConstant.Unauthorized, err).Send(c)
 	}
 	result, err := h.service.UpdateById(c.Request().Context(), payload)
 	if err != nil {
@@ -99,20 +95,17 @@ func (h *handler) UpdateById(c echo.Context) error {
 }
 
 func (h *handler) DeleteById(c echo.Context) error {
-	authHeader := c.Request().Header.Get("Authorization")
-	jwtClaims, err := util.ParseJWTToken(authHeader)
-	if err != nil {
-		return res.ErrorBuilder(&res.ErrorConstant.Unauthorized, err).Send(c)
-	}
-
-	log.Println(jwtClaims)
-
 	payload := new(pkgdto.ByIDRequest)
 	if err := c.Bind(payload); err != nil {
 		return res.ErrorBuilder(&res.ErrorConstant.BadRequest, err).Send(c)
 	}
 	if err := c.Validate(payload); err != nil {
 		return res.ErrorBuilder(&res.ErrorConstant.Validation, err).Send(c)
+	}
+	authHeader := c.Request().Header.Get("Authorization")
+	jwtClaims, err := util.ParseJWTToken(authHeader)
+	if (err != nil) || (jwtClaims.UserID != payload.ID) || (jwtClaims.RoleID != uint(enum.Admin)) {
+		return res.ErrorBuilder(&res.ErrorConstant.Unauthorized, err).Send(c)
 	}
 	result, err := h.service.DeleteById(c.Request().Context(), payload)
 	if err != nil {
